// example/src/theme/styled.d.ts
import { FlattenSimpleInterpolation, ThemedCssFunction } from 'styled-components'

export type Color = string
export interface Colors {
  // GUNUI POOL
  red4: Color
  red5: Color
  heading3: Color
  heading6: Color

  border: Color
  backIcon: Color
  bgHover: Color
  borderHover: Color
  iconHover: Color

  // base
  white: Color
  black: Color

  // text
  text1: Color
  text2: Color
  text3: Color
  text4: Color
  text5: Color
  text6: Color
  text7: Color
  text8: Color
  text9: Color
  text10: Color
  text11: Color
  text12: Color
  text13: Color

  // backgrounds / greys
  bg0: Color
  bg1: Color
  bg2: Color
  bg3: Color
  bg4: Color
  bg5: Color
  bg6: Color
  bg7: Color
  bg8: Color
  bg9: Color
  bg10: Color
  bg11: Color
  bg12: Color
  bg13: Color
  bg14: Color

  border1: Color

  modalBG: Color
  advancedBG: Color

  //blues
  primary1: Color
  primary2: Color
  primary3: Color
  primary4: Color
  primary5: Color

  primaryText1: Color

  // pinks
  secondary1: Color
  secondary2: Color
  secondary3: Color

  // other
  red1: Color
  red2: Color
  red3: Color
  red4: Color
  red5: Color
  red6: Color

  green1: Color

  yellow1: Color
  yellow2: Color
  yellow3: Color

  blue1: Color
  blue2: Color
  blue3: Color
  blue4: Color

  error: Color
  success: Color
  warning: Color

  progressBarBackgroundColor: Color
  progressBarBorderColor1: Color
  progressBarBorderColor2: Color

  tabTextStyle: string

  dropdownShadow: string
  dropdownItemTextColor: Color
  dropdownItemHoverBackground: Color
  dropdownItemHoverBorder: Color
}

export interface Grids {
  sm: number
  md: number
  lg: number
}

declare module 'styled-components' {
  export interface DefaultTheme extends Colors {
    grids: Grids

    // shadows
    shadow1: string

    // media queries
    mediaWidth: {
      upToExtraSmall: ThemedCssFunction<DefaultTheme>
      upToSmall: ThemedCssFunction<DefaultTheme>
      upToMedium: ThemedCssFunction<DefaultTheme>
      upToLarge: ThemedCssFunction<DefaultTheme>
    }

    // css snippets
    flexColumnNoWrap: FlattenSimpleInterpolation
    flexRowNoWrap: FlattenSimpleInterpolation
  }
}